split character skeleton functionality into its own script
